pyproject.toml
[project]
name = "brief-my-trade"
version = "0.1.0"
description = "메리츠증권 매매기록 텔레그램 봇"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "yfinance>=0.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brief_my_trade"]
